+++ lib/src/io.dart
@@ -67,8 +67,10 @@ bool processDirectory(FormatterOptions options, Directory directory) {
 bool processFile(FormatterOptions options, File file, {String label}) {
   if (label == null) label = file.path;
 
-  var formatter =
-      new DartFormatter(indent: options.indent, pageWidth: options.pageWidth);
+  var formatter = new DartFormatter(
+      indent: options.indent,
+      pageWidth: options.pageWidth,
+      fixes: options.fixes);
   try {
     var source = new SourceCode(file.readAsStringSync(), uri: file.path);
     options.reporter.beforeFile(file, label);