+++ test/formatter_test.dart
@@ -5,13 +5,11 @@
 @TestOn("vm")
 library dart_style.test.formatter_test;
 
-import 'dart:io';
-import 'dart:mirrors';
-
-import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
 
 import 'package:dart_style/dart_style.dart';
+
+import 'utils.dart';
 
 void main() {
   testDirectory("comments");
@@ -137,119 +135,4 @@ void main() {
     expect(() => formatter.format("var i = 1;"),
         throwsA(new isInstanceOf<UnexpectedOutputException>()));
   });
-}
-
-/// Run tests defined in "*.unit" and "*.stmt" files inside directory [name].
-void testDirectory(String name) {
-  var indentPattern = new RegExp(r"^\(indent (\d+)\)\s*");
-
-  // Locate the "test" directory. Use mirrors so that this works with the test
-  // package, which loads this suite into an isolate.
-  var testDir = p.dirname(currentMirrorSystem()
-      .findLibrary(#dart_style.test.formatter_test)
-      .uri
-      .toFilePath());
-
-  var entries = new Directory(p.join(testDir, name))
-      .listSync(recursive: true, followLinks: false);
-  for (var entry in entries) {
-    if (!entry.path.endsWith(".stmt") && !entry.path.endsWith(".unit")) {
-      continue;
-    }
-
-    group("$name ${p.basename(entry.path)}", () {
-      // Explicitly create a File, in case the entry is a Link.
-      var lines = new File(entry.path).readAsLinesSync();
-
-      // The first line may have a "|" to indicate the page width.
-      var pageWidth;
-      if (lines[0].endsWith("|")) {
-        pageWidth = lines[0].indexOf("|");
-        lines = lines.skip(1).toList();
-      }
-
-      var i = 0;
-      while (i < lines.length) {
-        var description = lines[i++].replaceAll(">>>", "").trim();
-
-        // Let the test specify a leading indentation. This is handy for
-        // regression tests which often come from a chunk of nested code.
-        var leadingIndent = 0;
-        var indentMatch = indentPattern.firstMatch(description);
-        if (indentMatch != null) {
-          leadingIndent = int.parse(indentMatch[1]);
-          description = description.substring(indentMatch.end);
-        }
-
-        if (description == "") {
-          description = "line ${i + 1}";
-        } else {
-          description = "line ${i + 1}: $description";
-        }
-
-        var input = "";
-        while (!lines[i].startsWith("<<<")) {
-          input += lines[i++] + "\n";
-        }
-
-        var expectedOutput = "";
-        while (++i < lines.length && !lines[i].startsWith(">>>")) {
-          expectedOutput += lines[i] + "\n";
-        }
-
-        // TODO(rnystrom): Stop skipping these tests when possible.
-        if (description.contains("(skip:")) {
-          print("skipping $description");
-          continue;
-        }
-
-        test(description, () {
-          var isCompilationUnit = p.extension(entry.path) == ".unit";
-
-          var inputCode =
-              _extractSelection(input, isCompilationUnit: isCompilationUnit);
-
-          var expected = _extractSelection(expectedOutput,
-              isCompilationUnit: isCompilationUnit);
-
-          var formatter =
-              new DartFormatter(pageWidth: pageWidth, indent: leadingIndent);
-
-          var actual = formatter.formatSource(inputCode);
-
-          // The test files always put a newline at the end of the expectation.
-          // Statements from the formatter (correctly) don't have that, so add
-          // one to line up with the expected result.
-          var actualText = actual.text;
-          if (!isCompilationUnit) actualText += "\n";
-
-          // Fail with an explicit message because it's easier to read than
-          // the matcher output.
-          if (actualText != expected.text) {
-            fail("Formatting did not match expectation. Expected:\n"
-                "${expected.text}\nActual:\n$actualText");
-          }
-
-          expect(actual.selectionStart, equals(expected.selectionStart));
-          expect(actual.selectionLength, equals(expected.selectionLength));
-        });
-      }
-    });
-  }
-}
-
-/// Given a source string that contains ‹ and › to indicate a selection, returns
-/// a [SourceCode] with the text (with the selection markers removed) and the
-/// correct selection range.
-SourceCode _extractSelection(String source, {bool isCompilationUnit: false}) {
-  var start = source.indexOf("‹");
-  source = source.replaceAll("‹", "");
-
-  var end = source.indexOf("›");
-  source = source.replaceAll("›", "");
-
-  return new SourceCode(source,
-      isCompilationUnit: isCompilationUnit,
-      selectionStart: start == -1 ? null : start,
-      selectionLength: end == -1 ? null : end - start);
 }